refactor: customer matching and business logic from new classes
