+++ java/src/main/java/codingdojo/CustomerMatchMaker.java
@@ -1,0 +1,75 @@
+package codingdojo;
+
+public class CustomerMatchMaker {
+    private final CustomerDataAccess customerDataAccess;
+
+    public CustomerMatchMaker(CustomerDataAccess customerDataAccess) {
+        this.customerDataAccess = customerDataAccess;
+    }
+
+    public CustomerMatches loadCustomerMatches(ExternalCustomer externalCustomer) {
+        if (externalCustomer.isCompany()) {
+            return loadCompany(externalCustomer);
+        } else {
+            return loadPerson(externalCustomer);
+        }
+    }
+
+    private CustomerMatches loadCompany(ExternalCustomer externalCustomer) {
+
+        final String externalId = externalCustomer.getExternalId();
+        final String companyNumber = externalCustomer.getCompanyNumber();
+
+        CustomerMatches customerMatches = customerDataAccess.loadCompanyCustomer(externalId, companyNumber);
+
+        if (customerMatches.getCustomer() != null
+                && !CustomerType.COMPANY.equals(customerMatches.getCustomer().getCustomerType())) {
+            throw new ConflictException(
+                    "Existing customer for externalCustomer " + externalId + " already exists and is not a company");
+        }
+
+        if ("ExternalId".equals(customerMatches.getMatchTerm())) {
+            String customerCompanyNumber = customerMatches.getCustomer().getCompanyNumber();
+            if (!companyNumber.equals(customerCompanyNumber)) {
+                customerMatches.getCustomer().setMasterExternalId(null);
+                customerMatches.addDuplicate(customerMatches.getCustomer());
+                customerMatches.setCustomer(null);
+                customerMatches.setMatchTerm(null);
+            }
+        } else if ("CompanyNumber".equals(customerMatches.getMatchTerm())) {
+            String customerExternalId = customerMatches.getCustomer().getExternalId();
+            if (customerExternalId != null && !externalId.equals(customerExternalId)) {
+                throw new ConflictException("Existing customer for externalCustomer " + companyNumber
+                        + " doesn't match external id " + externalId + " instead found " + customerExternalId);
+            }
+            Customer customer = customerMatches.getCustomer();
+            customer.setExternalId(externalId);
+            customer.setMasterExternalId(externalId);
+            customerMatches.addDuplicate(null);
+        }
+
+        return customerMatches;
+    }
+
+    private CustomerMatches loadPerson(ExternalCustomer externalCustomer) {
+        final String externalId = externalCustomer.getExternalId();
+
+        CustomerMatches customerMatches = customerDataAccess.loadPersonCustomer(externalId);
+
+        if (customerMatches.getCustomer() != null) {
+            if (!CustomerType.PERSON.equals(customerMatches.getCustomer().getCustomerType())) {
+                throw new ConflictException(
+                        "Existing customer for externalCustomer " + externalId + " already exists and is not a person");
+            }
+
+            if (!"ExternalId".equals(customerMatches.getMatchTerm())) {
+                Customer customer = customerMatches.getCustomer();
+                customer.setExternalId(externalId);
+                customer.setMasterExternalId(externalId);
+            }
+        }
+
+        return customerMatches;
+    }
+
+}
